Fixed radar Chart Bug
I added a sixth value to the list of results on the radar chart because it doesn't work properly with just 5 values.
Not sure why, but it works now, so we can just do it this way for now until we run into another problem.
Maybe the original Radar Chart script ignores the 5th value, maybe to finish the pentagon shape it is trying to make it needs two values for the "top" point of the pentagon. We can play around with the values to find out.

Instead of making the lowest result of the test 0, I changed it so that the lowest result of the test would be 2.
The reason we do it this way is because when the test is done, some schools might have a negative value, which doesn't work for what we want. So we take all the results of the test for each school and add the lowest number to all of them as its absolute value.

To give an example, if someone gets the scores:
(K , 20) (VE , 16) (NL , 5) (R , -5) (U , -13)
then we will take the lowest score: (U , -13),
and take its absolute value: (13)
and add that to all the scores so that the scores become:
--> (K , 33) (VE , 29) (NL , 18) (R , 8) (U , 0)
This distorts the ratios between the scores a bit, but is better than negative values.

We found now that 0 is not great because it has problems showing up on the radar chart and breaks other values,
so instead of zero, the lowest value now is going to be 2, just in case 1 is not clear enough.

We can try different solutions when we encounter another problem.